Cleanup, address comments
